Repackage to use resourceserver
- Changed base package to o.s.s.oauth2.resourceserver

- Changed samples to o.s.s.samples.config

Fixes: gh-1
