+++ samples/boot/oauth2/resource-server/pom.xml
