+++ samples/boot/oauth2/resource-server/src/main/resources/data.sql
