+++ samples/boot/oauth2/resource-server/src/main/resources/id_rsa_pkcs8.pub
