--- samples/boot/oauth2/src/main/java/org/springframework/security/samples/oauth2/rs/auth0/MessageController.java
@@ -1,48 +1,0 @@
-/*
- * Copyright 2002-2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.security.samples.oauth2.rs.auth0;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import javax.validation.Valid;
-
-/**
- * @author Rob Winch
- */
-@RestController
-@RequestMapping("/messages")
-public class MessageController {
-	private final MessageRepository messages;
-
-	public MessageController(MessageRepository messages) {
-		this.messages = messages;
-	}
-
-	@GetMapping("/{id}")
-	public Message findOne(@PathVariable Long id) {
-		return this.messages.findById(id).orElse(null);
-	}
-
-	@PostMapping
-	public Message save(@Valid @RequestBody Message message) {
-		return this.messages.save(message);
-	}
-}
