--- samples/boot/oauth2/src/main/java/org/springframework/security/samples/oauth2/rs/auth0/MessageRepository.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2002-2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.security.samples.oauth2.rs.auth0;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.security.access.prepost.PreAuthorize;
-
-import java.util.Optional;
-
-/**
- * @author Rob Winch
- */
-public interface MessageRepository extends CrudRepository<Message,Long> {
-    @PreAuthorize("authentication.hasClaim('scope').thatMatches('.*message.read.*')")
-    Optional<Message> findById(Long id);
-
-    @PreAuthorize("hasAuthority('message.write')")
-    Message save(Message message);
-}
