--- samples/boot/oauth2/src/test/java/org/springframework/security/samples/oauth2/rs/auth0/PublicKeyOnlyRSAKeyProviderTests.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2002-2018 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.security.samples.oauth2.rs.auth0;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.runners.MockitoJUnitRunner;
-
-import java.security.interfaces.RSAPublicKey;
-
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-public class PublicKeyOnlyRSAKeyProviderTests {
-	PublicKeyOnlyRSAKeyProvider provider = (s) -> null;
-
-	@Test
-	public void whenTryPrivateKey_thenError() {
-		assertThatThrownBy(() -> this.provider.getPrivateKey())
-			.isInstanceOf(UnsupportedOperationException.class);
-	}
-
-	@Test
-	public void whenTryPrivateKeyId_thenError() {
-		assertThatThrownBy(() -> this.provider.getPrivateKeyId())
-			.isInstanceOf(UnsupportedOperationException.class);
-	}
-}
